requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
openai==1.12.0
langchain-openai==0.0.5
langchain-chroma==0.1.2
langchain==0.1.6
langchain-core==0.1.23
langchain-community==0.0.20
chromadb==0.4.24
google-genai==0.3.5
Pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
